// webui/src/features/systemConfig/types.ts
export type RuntimeConfig = {
  request_log_enabled: boolean;
  reverse_proxy_log_detail_enabled: boolean;
  reverse_proxy_log_req_headers_max_bytes: number;
  reverse_proxy_log_req_body_max_bytes: number;
  reverse_proxy_log_resp_headers_max_bytes: number;
  reverse_proxy_log_resp_body_max_bytes: number;
  max_consecutive_failures: number;
  max_latency_test_interval: string;
  max_authority_latency_test_interval: string;
  max_egress_test_interval: string;
  latency_test_url: string;
  latency_authorities: string[];
  p2c_latency_window: string;
  latency_decay_window: string;
  cache_flush_interval: string;
  cache_flush_dirty_threshold: number;
};

export type EnvConfig = {
  cache_dir: string;
  state_dir: string;
  log_dir: string;
  listen_address: string;
  resin_port: number;
  api_max_body_bytes: number;
  max_latency_table_entries: number;
  probe_concurrency: number;
  geoip_update_schedule: string;
  default_platform_sticky_ttl: string;
  default_platform_regex_filters: string[] | null;
  default_platform_region_filters: string[] | null;
  default_platform_reverse_proxy_miss_action: string;
  default_platform_reverse_proxy_empty_account_behavior: string;
  default_platform_reverse_proxy_fixed_account_header: string;
  default_platform_allocation_policy: string;
  probe_timeout: string;
  resource_fetch_timeout: string;
  proxy_transport_max_idle_conns: number;
  proxy_transport_max_idle_conns_per_host: number;
  proxy_transport_idle_conn_timeout: string;
  request_log_queue_size: number;
  request_log_queue_flush_batch_size: number;
  request_log_queue_flush_interval: string;
  request_log_db_max_mb: number;
  request_log_db_retain_count: number;
  metric_throughput_interval_seconds: number;
  metric_throughput_retention_seconds: number;
  metric_bucket_seconds: number;
  metric_connections_interval_seconds: number;
  metric_connections_retention_seconds: number;
  metric_leases_interval_seconds: number;
  metric_leases_retention_seconds: number;
  metric_latency_bin_width_ms: number;
  metric_latency_bin_overflow_ms: number;
  admin_token_set: boolean;
  proxy_token_set: boolean;
  admin_token_weak: boolean;
  proxy_token_weak: boolean;
};

export type RuntimeConfigPatch = Partial<RuntimeConfig>;